Windows: Refactor daemon
